UART bug fix.  Updated how the uart is opened.
In docker, if the pthreads library was linked to the code (i.e., in
cmake Threads::Threads)  the uart would not receive data from the uart
correctly, instead it would time out.

I removed a usleep from uart_open (which likely should not have been
there.  comments said it was needed to make sure that we flushed the
buffer when using a usb serial port, but the code does not depend on the
initial buffer being flushed, and I htink it was a bad idea to have a
sleep of some duration.  My hunch is that this was causing the
weird behavior in the docker/pthreads combination.   I also now
initialize the termios structure with settings I read from the system.
This might also actually help. Anyway, the bug seems to be fixed for
now.