Add new protocol module for BASE
This commit (1) renames the "omni" module to "base" so that it's more appropriately named,
and (2) adds a new protocol module for controlling the Base Tiva board. The current module implements test functions to change the colors of an LED and to send a twist from a PC to the Tiva using the packet_console.cpp file.