Robustified the error when the estop is pressed.
When estop is pressed while delta is streaming commands, previously
the JC would go into error mode and stop responding to UC-1.

UC-1 would then timeout.  But it would also detect the estop, causing
a double error.

Now, when setting core_error, the uart is flushed of its next packet.
If more than 0 bytes of data were flushed, the reciever sends an error
command, to notify the sender that an error occurred.   If 0 bytes of
data were flushed, then the error did not occur midway through a
communication and no unsolicited error byte is sent.