Update odometry calculation code, add unit tests
Odometry has now been corrected for the omni robot. Calculations have been added to determine the change in position relative to the last known configuration in the body frame, and then transform it to the space frame. The change is then multiplied by a time step and added to the existing robot pose to update its position and orientation in the space frame. Unit tests have also been added to test for translational and rotational movement.