Change the soft max velocity parameter
The parameter is now 1.25% of the desired soft maximum of 2Pi (60rpm). The percentage is used to create an absolute maximum used for a emergency stop. We would like to have the robot to maintain normal operation even if the speed of the wheels goes just slightly above the desired soft max.