Working on publishing knee data.  Created a stub in delta_robot.c
to compute the knee angles from the robot state.  Need to generate tests
for this as well.

Working towards using robot_state_publisher to use the urdf to publish tf frames,
based on the actual robot position.

Also, made DELTA_ROBOT a single constant that is stored in delta_robot.c
THis stores the default parameters for the robot, and will amtch the
physical robot.

These parameters are now ultimately defined in common/include/common/parameters.cmake

so that they 1. do not depend on ROS code. 2. Are available to ROS code at build time.