Minor optimization to writing a pin.
The original code was slightly inefficient because it used the mask of which pin
both in the address and in what values to set.  By only using the mask in the address
we can simply write all 1 to the bit address if we want to write a 1 to a single pin
or all 0 to the bit address if we want to write a 0: the mask causes the register
to ignore the irrelevant bits.

This should technically make our code faster than GPIOPinWrite when manipulating a single
bit.