Slightly improve pin_read performance by not using GPIOPinRead and ANDing result.
The next step would be to change the bit packing of the pin uint32_t to be
incompatible with driverlib/pin_map.h.   Instead the pin value would be
directly the offset into the HWREG.  Implementing this would require
some deeper changes, so I will leave it for now.

We could also inline the pin_write and pin_read functions