Add error checking code to pin_configure.
When pins are configured the port and number that are extracted are checked for
validity.  This makes the code a bit safer from logic errors now
that port and number are no longer checked due to performance reasons:
These values are now checked when the pin is configured for the first time.
So as long as you are using configured pins, there should be no problem
(and if you aren't other errors will occur).