Made nuhal_install version checking ignore architecture when there were interface sources.
The nuhal_all library is distributed in source code form so it is independent of architecture.
nuhal_install now checks for this automatically and find_package won't verify architecture in this case.

This could be problematic for libraries that have pre-compiled components and soruce code distributed,
but then you cna just not use nuhal_install