Updated the install process so that now each platform-specific library is called nuhal.
This is instead of having nuhal_linux and nuhal_tiva.  It is the quickest and simplest way
to let client code simply include nuhal::nuhal instead of specifying the platform.

An alternative would be for each library to install as nuhal_tiva and nuhal_linux respectively
and then, one of the packages (perhaps nuhal_all) creates a nuhal-config.cmake.

nuhal-config.cmake would then have logic to select the proper library based on the platform.
This gives some additional flexibility; however, it is more maitnentece