Changed some dependencies and public/private target_link_libraries for nuhal.
Nuhal tiva now includes startup code automatically.
nuhal_all is now a private dependency, to prevent the code from being re-compiled by downstream dependees
nuhal_flags is now a public dependency of nuhal.  This means that each dependee
of nuhal automatically uses the same compile flags.  These flags are a bit opinionated, but they
are the flags I use in most of my projects.

If need be, down the line, I could export a version that does not have the flags public.