Begin formalizing how the cross-platform core part
 of the library is made.

Each cross-platform library consists of two parts

1. The cross-platform part: this code is shared by all platforms
2. The platform-specific part

This is setup so that the cross-platform part and platform specific
parts can all be in separate directories with separate cmakelists.txt
(and therefore be separate catkin_packages when using catkin).

When the cross-platform part is "compiled" what is really happening is
that source files and other information is being gathered into a cmake
target.  However, compilation is deferred until we build the
platform-specific library.

The cross-platform library therefore consists of two parts
1. Source code files.  These are part of its interface, so the
   platform-specific library target_link_libraries against this and then
    when it is compiled it compiles those source files (with the
    appropriate cross-compiling toolchain).
    The platform-specific library should take these files in as PRIVATE
    because its dependees should not recompile the files.
    Thus this is in the _private part of the library
2. Include files and compiler flags. These must be passed on to users
    of the platform-specific libraries so these are  _public.

The library that users actually uses links against both the PRIVATE and
PUBLIC pieces of the core library.